* import.c (expand_at_signs): Typecasting for fwrite results. * log-buffer.c (log_buffer_input, log_buffer_output, log_buffer_initialize): Ditto.
* log-buffer.c (log_buffer_initialize): Protect reference to
fatal_errors.

* parseinfo.c (parse_config): Protect reference to PrimaryServer
using #ifdef PROXY_SUPPORT.

* ls.c (ls_fileproc): Remove unused variables.
* subr.c (increment_revnum): Ditto.
* vers_ts.c (time_stamp): Ditto.

CVS-ID: src/ChangeLog 1.2873
CVS-ID: src/import.c 1.151
CVS-ID: src/log-buffer.c 1.12
CVS-ID: src/ls.c 1.14
CVS-ID: src/parseinfo.c 1.56
CVS-ID: src/subr.c 1.119
CVS-ID: src/vers_ts.c 1.58

+++ src/import.c
@@ -1624,7 +1624,7 @@ expand_at_signs (const char *buf, off_t size, FILE *fp)
 	size -= len;
     }
 
-    if (fwrite (cp, 1, size, fp) != size)
+    if (fwrite (cp, 1, size, fp) != (size_t) size)
 	return EOF;
 
     return 1;
